common: remove solana web3 dep
